opencv: checking existence of /usr/lib
Remove install errors:

  mv: cannot stat '.../tmp/work/mips64-wrs-linux/opencv/3.1+gitAUTOINC
  +92387b1ef8-r0/image/usr/lib/*': No such file or directory

(cherry picked from commit cefba87ff0a5f12cdb377038d67ab8550d500e82)

+++ meta-oe/recipes-support/opencv/opencv_3.1.bb
@@ -138,7 +138,7 @@ do_install_append() {
     sed -i '/blobtrack/d' ${D}${includedir}/opencv/cvaux.h
 
     # Move Python files into correct library folder (for multilib build)
-    if [ "$libdir" != "/usr/lib" ]; then
+    if [ "$libdir" != "/usr/lib" -a -d ${D}/usr/lib ]; then
         mv ${D}/usr/lib/* ${D}/${libdir}/
         rm -rf ${D}/usr/lib
     fi
